No refresh publication function

+++ backend/controllers/publication.js
@@ -8,11 +8,12 @@ const fs = require("fs");
 
 // Import Publication class
 
-let publication = require("../models/Publication")
+let publication = require("../models/Publication");
+const { json } = require("express");
 
 // Créer une publication
 
-exports.createPublication = (req, res, next) => {
+exports.createPublication = (req, response, next) => {
   let filename;
   if(req.body.image){
     filename = null
@@ -28,6 +29,20 @@ exports.createPublication = (req, res, next) => {
     if(err) throw err;
     console.log(res);
     console.log("Publication crée");
+
+    let sql2 = `SELECT MAX(id) FROM postes WHERE nom = "${req.body.nom}" AND prenom = "${req.body.prenom}" AND content = "${req.body.content}" AND user_id = "${req.body.user_id}"`;
+    let query2 = connexion.query(sql2, (err,res) => {
+      if(err) throw err;
+      data = JSON.parse(JSON.stringify(res))[0];
+      post_id = data[Object.keys(data)[0]];
+      console.log(post_id)
+
+      let sql3 = `SELECT * FROM postes WHERE id = "${post_id}"`
+      let query3 = connexion.query(sql3, (err, res) => {
+        if(err) throw err;
+        return response.status(200).json(res[0]);
+      })
+    })
   });
 };
 
@@ -48,7 +63,7 @@ exports.getAllPublication = (req,response, next) => {
 
 // Modification d'une publication
 
-exports.modifyPublication = (req,res, next) => {
+exports.modifyPublication = (req,response, next) => {
 
   if(req.body.id === req.body.user_id || req.body.role === "Admin"){
     let sql = `UPDATE postes SET content = '${req.body.content}' WHERE id = ${req.params.id}`;
@@ -60,26 +75,32 @@ exports.modifyPublication = (req,res, next) => {
   
     if(!req.body.image){
       let sql2 = `UPDATE postes SET file = '${req.protocol}://${req.get('host')}/images/${req.file.filename}' WHERE id = ${req.params.id}`;
-      let query = connexion.query(sql2, (err, res) => {
+      let query2 = connexion.query(sql2, (err, res) => {
       if(err) throw err;
       console.log(res);
       console.log("Publication Image modifié");
-      const filename = req.body.filename.split('http://localhost:3000/images/')[1];
-      fs.unlink(`images/${filename}`, (err) => {
+      if(req.body.filename){
+        const filename = req.body.filename.split('http://localhost:3000/images/')[1];
+        fs.unlink(`images/${filename}`, (err) => {
         if(err) throw err;
         console.log("file unlinked")
       })
+      }
     });
     }
+    let sql3 = `SELECT * FROM postes WHERE id = "${req.params.id}"`
+    let query3 = connexion.query(sql3, (err, res) => {
+      return response.status(200).json(res[0])
+    })
   }
 };
 
 // Supprimer une publication
 
-exports.deletePublication = (req,res, next) =>{
+exports.deletePublication = (req,response, next) =>{
 
   if(req.body.id === req.body.user_id || req.body.role === "Admin"){
-    let sql = `DELETE FROM postes WHERE id = ${req.params.id /*req.body.post_delete.post_id*/}`
+    let sql = `DELETE FROM postes WHERE id = ${req.params.id}`
     let query = connexion.query(sql, (err, res) => {
       if(err) throw err;
       console.log(res);
@@ -92,6 +113,7 @@ exports.deletePublication = (req,res, next) =>{
           console.log("file unlinked")
         })
       }
+      return response.status(200).json(req.params.id)
     })
   }
 }
